schedule.html: adjust lunch break and finish time
I moved the lunch break to 12:30-13:30 and the finish time to 5pm as requested by Anna-Lisa Döring mail 17 Nov 2023. The slot for wrap-ups is 30' and, on the last day, the post-workshop survey is 10' according to the previous schedule. The instructors may further customize a schedule of the individual lessons including breaks.